src/receipt/ocr.py
def perform_ocr(image_path):
    import pytesseract
    from PIL import Image

    # Load the image from the specified path
    image = Image.open(image_path)

    # Use pytesseract to perform OCR on the image
    text = pytesseract.image_to_string(image)

    return text

def extract_receipt_data(ocr_text):
    # Placeholder function to extract relevant data from OCR text
    # This function should be implemented to parse the text and extract
    # items, prices, and other relevant information from the receipt.
    receipt_data = {}
    # Example parsing logic (to be implemented)
    lines = ocr_text.split('\n')
    for line in lines:
        # Implement logic to extract data from each line
        pass

    return receipt_data